% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{CellCycleAnalysis}
\alias{CellCycleAnalysis}
\title{CellCycleAnalysis}
\usage{
CellCycleAnalysis(SeuratObject, savePath = NULL, geneSets = NULL, ...)
}
\arguments{
\item{SeuratObject}{A Seurat object}

\item{savePath}{Path to save}

\item{geneSets}{Gene sets for cell cycle analysis}
}
\value{
A Seurat object after NMF
}
\description{
calculate the score of cell cycle
}
