% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{cellTypeScore}
\alias{cellTypeScore}
\title{cellTypeScore}
\usage{
cellTypeScore(
  SeuratObject,
  savePath = NULL,
  geneSets = NULL,
  select = "main",
  assay = "SCT",
  method = "average",
  species = "human",
  total.image.name = "cellTypeScore.png",
  ...
)
}
\arguments{
\item{SeuratObject}{A Seurat object}

\item{savePath}{}
}
\value{
A Seurat object after scoring
}
\description{
calculate the score of some cell types
}
