% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clusterBarPlot}
\alias{clusterBarPlot}
\title{clusterBarPlot}
\usage{
clusterBarPlot(
  spot.annotation,
  sel.col = "spot.Type",
  spot.colors = NULL,
  legend.title = NULL,
  legend.position = "bottom"
)
}
\arguments{
\item{spot.manifest}{}
}
\value{

}
\description{
Draw
}
