% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{EMTAnalysis}
\alias{EMTAnalysis}
\title{EMTAnalysis}
\usage{
EMTAnalysis(SeuratObject, savePath = NULL, geneSets = NULL, ...)
}
\arguments{
\item{SeuratObject}{A Seurat object}

\item{savePath}{Path to save}

\item{geneSets}{Gene sets}
}
\value{
A Seurat object after NMF
}
\description{
calculate the score of EMT
}
