% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Feature_Plot}
\alias{Feature_Plot}
\title{Feature_Plot}
\usage{
Feature_Plot(
  object,
  features,
  legend.name = NULL,
  legend.color = getDefaultFeatureColors(type = "seq"),
  ...
)
}
\value{
plots
}
\description{
Draw continuous features
}
