% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{genePropPlot}
\alias{genePropPlot}
\title{genePropPlot}
\usage{
genePropPlot(gene.manifest, expr.frac)
}
\arguments{
\item{gene.manifest}{A dataframe of gene information}

\item{expr.frac}{A matrix including the expression fraction of each gene}
}
\value{
plots
}
\description{
Drow gene proportion
}
