% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GeneSpatialPlot}
\alias{GeneSpatialPlot}
\title{GeneSpatialPlot}
\usage{
GeneSpatialPlot(object, gene, ...)
}
\arguments{
\item{object}{Seurat object}

\item{gene}{gene in object}
}
\value{

}
\description{
Plot gene expression level in spatial
}
