% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_spatial}
\alias{geom_spatial}
\title{geom_spatial}
\usage{
geom_spatial(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\value{

}
\description{
Draw spatial figure
}
