% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDefaultFeatureColors}
\alias{getDefaultFeatureColors}
\title{getDefaultFeatureColors}
\usage{
getDefaultFeatureColors(type = "seq", name = NULL)
}
\description{
Return the colors for continuous features
}
