% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGeneManifest}
\alias{getGeneManifest}
\title{getGeneManifest}
\usage{
getGeneManifest(data.path, h5 = FALSE, filter.matrix = TRUE)
}
\value{
A dataframe of gene information
}
\description{
Read genes.tsv
}
