% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMouseGene}
\alias{getMouseGene}
\title{#' filePathCheck
#'
#' check the format of the file path
#'
#' @param filePath
#'
#' @return filePath
#'
#'
filePathCheck <- function(filePath){
    return(gsub("\\\\", "/", filePath))
}
getMouseGene}
\usage{
getMouseGene(hg.genes)
}
\description{
#' filePathCheck
#'
#' check the format of the file path
#'
#' @param filePath
#'
#' @return filePath
#'
#'
filePathCheck <- function(filePath){
    return(gsub("\\\\", "/", filePath))
}
getMouseGene
}
