% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{grid_arrange_shared_legend}
\usage{
grid_arrange_shared_legend(
  ...,
  all.p,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right")
)
}
\value{
plots
}
\description{
grid arrange
}
