% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{histPlot.data.frame}
\alias{histPlot.data.frame}
\title{histPlot.data.frame}
\usage{
histPlot.data.frame(
  data,
  value,
  color = "#a788ab",
  xlines = c(),
  y.label = "Spot number"
)
}
\value{

}
\description{
Draw histogram of data.frame
}
