% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limitData}
\alias{limitData}
\title{limitData}
\usage{
limitData(data, min = NULL, max = NULL)
}
\arguments{
\item{data}{data}

\item{min}{min}

\item{max}{max}
}
\value{
A new data
}
\description{
limit data between min and max
}
