% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stStatistics.R
\name{Load_Spatial}
\alias{Load_Spatial}
\title{Load_Spatial}
\usage{
Load_Spatial(
  data.dir,
  assay = "Spatial",
  slice = "slice1",
  sampleName = "SeuratProject",
  rm.isolated = TRUE,
  region.threshold = 3,
  h5 = FALSE,
  filter.matrix = TRUE,
  to.upper = FALSE,
  ...
)
}
\arguments{
\item{data.dir}{Directory containing the H5 file specified by filename and the image data in a subdirectory called spatial.
Under this path, folders 'filtered_feature_bc_matrix' and 'raw_feature_bc_matrix' exist generally.}

\item{assay}{Name of the initial assay}

\item{slice}{Name for the stored image of the tissue slice.}

\item{sampleName}{A character string giving a label for this sample.}

\item{rm.isolated}{A logical value indicating whether to remove isolated spots.}

\item{region.threshold}{A integer value, if the area of connected domain of spots are less than `region.threshold`,
they will be removed.}

\item{h5}{A logical value indicating whether to read h5 files, whose contents are the same as 'filtered_feature_bc_matrix' or 'raw_feature_bc_matrix'}

\item{filter.matrix}{Only keep spots that have been determined to be over tissue.}

\item{to.upper}{Converts all feature names to upper case. Can be useful when analyses require comparisons between human and mouse gene names for example.}

\item{...}{}
}
\value{
A Seurat object
}
\description{
Load a 10x Genomics Visium Spatial Experiment into a Seurat object (similar to Load10X_Spatial in Seurat)
}
