% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{plotMalignancy}
\alias{plotMalignancy}
\title{plotMalignancy}
\usage{
plotMalignancy(
  spot.annotation,
  coor.names = c("tSNE_1", "tSNE_2"),
  savePath = NULL
)
}
\value{
A plot list.
}
\description{
plotMalignancy
}
