% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pointDRPlot}
\alias{pointDRPlot}
\title{pointDRPlot}
\usage{
pointDRPlot(
  object,
  feature,
  reduction.func = "tsne",
  colors = NULL,
  title = NULL,
  discrete = T,
  limit.quantile = 0,
  limit.q.nz = FALSE,
  point.type = 1,
  point.size = NULL,
  base.size = 6,
  legend.position = "right",
  legend.title = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{}
}
\value{

}
\description{
plot features in a reduced dimensionality space
}
