% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{prepareCNV}
\alias{prepareCNV}
\title{prepareCNV}
\usage{
prepareCNV(
  SeuratObject,
  gene.manifest,
  assay = "Spatial",
  ref.object = NULL,
  species = "human",
  genome = "hg19",
  hg.mm.mix = F
)
}
\arguments{
\item{hg.mm.mix}{}
}
\value{

}
\description{
Get common genes in 'expr.data' and 'ref.data'
}
