% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stStatistics.R
\name{Read_Image}
\alias{Read_Image}
\title{Read_Image}
\usage{
Read_Image(
  image.dir,
  image.name = "tissue_lowres_image.png",
  filter.matrix = TRUE,
  rm.isolated = T,
  region.threshold = 3
)
}
\arguments{
\item{image.dir}{Path to directory with 10X Genomics visium image data; should include files tissue_lowres_iamge.png, scalefactors_json.json and tissue_positions_list.csv}

\item{filter.matrix}{Filter spot/feature matrix to only include spots that have been determined to be over tissue.}

\item{rm.isolated}{A logical value indicating whether to remove isolated spots.}

\item{region.threshold}{A integer value, if the area of connected domain of spots are less than `region.threshold`,
they will be removed.}
}
\value{
A VisiumV1 object
}
\description{
Load a 10x Genomics Visium Image (similar to Read10X_Image in Seurat)
}
