% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stStatistics.R
\name{rm_isolated_spot}
\alias{rm_isolated_spot}
\title{rm_isolated_spot}
\usage{
rm_isolated_spot(spm, region.threshold = 3)
}
\arguments{
\item{spm}{the data.frame of tissue.possition}

\item{region.threshold}{A integer value, if the area of connected domain of spots are less than `region.threshold`,
they will be removed.}
}
\value{
the data.frame of tissue.possition after removing isolated areas
}
\description{
Remove isolated spots
}
