% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{runCNVAnalysis}
\alias{runCNVAnalysis}
\title{runCNVAnalysis}
\usage{
runCNVAnalysis(
  SeuratObject,
  savePath,
  analysis.func = "copyKAT",
  ref.object = NULL,
  species = "human",
  ...
)
}
\arguments{
\item{SeuratObject}{A Seurat object}

\item{savePath}{A path to save the results files}

\item{analysis.func}{The function used to analyze CNV}

\item{ref.object}{A seuratObject used as the reference if using InferCNV}

\item{species}{"human" or "mouse", only human data can be used in CNV analysis by copyKAT}
}
\value{
A list containing a seuratObject, a list of results of intermediate process,
a list of plots and a logical value to show whether the process completed successfully
}
\description{
CNV analysis
}
