% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{runInferCNV}
\alias{runInferCNV}
\title{runInferCNV}
\usage{
runInferCNV(
  SeuratObject,
  savePath,
  ref.object,
  cluster.info = "seurat_clusters",
  species = "human",
  genome = "hg19"
)
}
\arguments{
\item{SeuratObject}{A seuratObejct to analyze CNV}

\item{savePath}{A path to store the results}

\item{ref.object}{A seuratObject used as the reference if using InferCNV}

\item{cluster.info}{The label to store cluster information in `SeuratObject@meta.data`}

\item{species}{"human" or "mouse"}

\item{genome}{}
}
\value{
A list containing a seuratObject, a list of results of intermediate process,
a list of plots and a logical value to show whether the process completed successfully
}
\description{
Analyze CNV using the method of InferCNV
}
