% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{runMalignancy}
\alias{runMalignancy}
\title{runMalignancy}
\usage{
runMalignancy(
  SeuratObject,
  savePath,
  assay = "Spatial",
  cutoff = 0.1,
  minspot = 3,
  p.value.cutoff = 0.5,
  coor.names = c("tSNE_1", "tSNE_2"),
  ref.object = NULL,
  species = "human",
  genome = "hg19",
  hg.mm.mix = F
)
}
\arguments{
\item{cutoff}{The cut-off for min average read counts per gene among
reference spots. The default is 0.1.}

\item{minspot}{An integer number used to filter gene. The default is 3.}

\item{p.value.cutoff}{The p-value to decide whether the distribution of
malignancy score is bimodality.}

\item{expr}{A Seurat object.}

\item{gene.manifest}{A data.frame of genes' manifest.}

\item{spot.annotation}{A data.frame of spots' annotation.}

\item{ref.data}{An expression matrix of gene by spot, which is used as the normal reference.
The default is NULL, and an immune spots or bone marrow spots expression matrix will be used for human or mouse species, respectively.}

\item{referAdjMat}{An adjacent matrix for the normal reference data.
The larger the value, the closer the spot pair is.
The default is NULL, and a SNN matrix of the default ref.data will be used.}
}
\value{
A list of cnvList, reference malignancy score, seurat object,
spot.annotatino, bimodal.pvalue, malign.thres, and all generated plots.
}
\description{
runMalignancy
}
