% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{runNMF}
\alias{runNMF}
\title{runNMF}
\usage{
runNMF(object, savePath, rank = 5, base.size = 8, pt.size = 1.6)
}
\arguments{
\item{object}{A Seurat object}

\item{savePath}{A path to save the results files(suggest to create a foler named by sample name)}

\item{rank}{A integer value to determine the dimension of decomposition matrix}
}
\value{
A list containing a seuratObject, a list of results of intermediate process,
a list of plots and a logical value to show whether the process completed successfully
}
\description{
perform NMF
}
