% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Spatial_Plot}
\alias{Spatial_Plot}
\title{Spatial_Plot}
\usage{
Spatial_Plot(
  object,
  feature,
  image.key = "slice1",
  show.tissue = F,
  crop = F,
  title = NULL,
  legend.title = NULL,
  legend.position = "right",
  legend.size = 3,
  colors = NULL,
  limit.quantile = 0,
  limit.type = "both",
  discrete = F,
  margin = 5,
  base.size = 8,
  pt.size = "auto",
  alpha = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{feature}{Feature in Seurat meta.data}
}
\value{

}
\description{
The most fundamental function
}
