% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SpatialDim_Plot}
\alias{SpatialDim_Plot}
\title{SpatialDim_Plot}
\usage{
SpatialDim_Plot(
  object,
  group.by = NULL,
  crop = TRUE,
  stroke = NA,
  legend.name = NULL,
  legend.color = NULL,
  ...
)
}
\value{
plots
}
\description{
Draw discrete features in the spatial location
}
