% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SpatialFeature_Plot}
\alias{SpatialFeature_Plot}
\title{SpatialFeature_Plot}
\usage{
SpatialFeature_Plot(
  object,
  features,
  crop = TRUE,
  stroke = NA,
  legend.name = NULL,
  legend.color = getDefaultFeatureColors(type = "seq"),
  ...
)
}
\value{
plots
}
\description{
Draw continuous features in the spatial location
}
