% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStructure.R
\name{SpatialStructureDetection}
\alias{SpatialStructureDetection}
\title{SpatialStructureDetection}
\usage{
SpatialStructureDetection(
  object,
  feature,
  slice = "slice1",
  global.Moran.I = 0.5,
  p.adjust.method = "fdr",
  p.thres.local = 0.05,
  p.adj.thres.local = 0.05,
  p.thres.z.score = 0.05,
  median.filter = TRUE,
  z.score.filter = TRUE
)
}
\arguments{
\item{object}{A Seurat object after stStatistics}

\item{feature}{The features (genes) of the spatial structure}

\item{slice}{Slice}

\item{global.Moran.I}{The threshold for global Morans' I. If the calculated global Morans' I is less than `global.Moran.I`,
it is considered that no corresponding spatial structure exists in the tissue}

\item{p.adjust.method}{The method used to adjust p values}

\item{p.thres.local}{The threshold for p values of local Morans' I}

\item{p.adj.thres.local}{The threshold for adjusted p values of local Morans' I}

\item{p.thres.z.score}{The threshold for z scores of local Morans' I}

\item{median.filter}{Whether to filter the results using the median of feature scores}

\item{z.score.filter}{Whether to filter the results based on z-scores}

\item{crop}{A logical value indicating whether to crop the image during plotting}

\item{CNV.ref.object}{A reference seurat object used in inferCNV}

\item{authorName}{A character string for authors name and will be shown in the report.}

\item{genReport}{A logical value indicating whether to generate a .html/.md report (suggest to set TRUE).}
}
\value{
A Seuratobject with results
}
\description{
Detect spatial structures in the tissue using global and local Morans' I.
}
