% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stAnnotation.R
\name{stAnnotation}
\alias{stAnnotation}
\title{stAnnotation}
\usage{
stAnnotation(
  object,
  savePath,
  rm.mito = T,
  rm.ribo = T,
  rm.ig = T,
  topGeneNum = 10,
  clst.resolution = 0.4,
  n.markers = 5,
  rank = 5,
  species = "human",
  cellTypeScore_fun = "average",
  cellTypeScore_select = "main",
  geneSets = NULL,
  analysis.func = "copyKAT",
  interaction.region.threshold = 20,
  crop = TRUE,
  verbose = F,
  authorName = NULL,
  genReport = T,
  bool.NMF = TRUE,
  bool.CellType = TRUE,
  bool.CNV = TRUE,
  bool.interaction = TRUE,
  bool.EMT = TRUE,
  bool.CellCycle = TRUE,
  bool.stem = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{savePath}{A path to save the results files(suggest to create a foler named by sample name).}

\item{rm.mito}{A logical value indicating whether to remove}

\item{crop}{A logical value indicating whether to crop the image during plotting}

\item{authorName}{A character string for authors name and will be shown in the report.}

\item{genReport}{A logical value indicating whether to generate a .html/.md report (suggest to set TRUE).}

\item{rm.isolated}{A logical value indicating whether to remove isolated spots}
}
\value{
A Seurat object after statistic
}
\description{
perform statistic
}
