% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{StemnessAnalysis}
\alias{StemnessAnalysis}
\title{StemnessAnalysis}
\usage{
StemnessAnalysis(SeuratObject, savePath = NULL, geneSets = NULL, ...)
}
\arguments{
\item{SeuratObject}{A seurat object to estimate stemness.}

\item{stem.sig}{An array of stemness signature. The default is NULL, and a prepared signature will be used.}
}
\value{

}
\description{
Estimate cell stemness according to the Spearman correlation with stemness signature.
}
