% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{TLSAnalysis}
\alias{TLSAnalysis}
\title{TLSAnalysis}
\usage{
TLSAnalysis(
  SeuratObject,
  savePath = NULL,
  geneSets = NULL,
  method = "AddModuleScore",
  crop = T,
  ...
)
}
\arguments{
\item{SeuratObject}{A Seurat object}

\item{savePath}{Path to save}

\item{geneSets}{Gene sets for TLS analysis}
}
\value{
A Seurat object after NMF
}
\description{
calculate the score of TLS
}
