% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{TumorCharacters}
\alias{TumorCharacters}
\title{TumorCharacters}
\usage{
TumorCharacters(
  object,
  savePath = NULL,
  species = "human",
  method = "AddModuleScore",
  crop = T
)
}
\arguments{
\item{object}{A Seurat object}

\item{savePath}{Path to save}

\item{species}{Species}

\item{method}{AddModuleScore or average}

\item{crop}{The parameter of plotting function}
}
\value{
A list including seurat oject, plots and results
}
\description{
calculate the score of tumor characters defined by cancerSEA
}
