% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{histPlot}
\alias{histPlot}
\title{histPlot}
\usage{
histPlot(
  object,
  feature,
  color = "#a788ab",
  xlines = c(),
  y.label = "Spot number"
)
}
\value{

}
\description{
Draw histogram of Seurat object
}
