% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{preDEheatmap}
\alias{preDEheatmap}
\title{preDEheatmap}
\usage{
preDEheatmap(
  object,
  col.name = "seurat_clusters",
  genes = NULL,
  spots = NULL,
  assay = NULL,
  slot = "scale.data",
  min.value = -2.5,
  max.value = 2.5
)
}
\arguments{
\item{object}{Seurat object}
}
\value{
plots
}
\description{
Do preparation for heatmap
}
