% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{prepareCNV}
\alias{prepareCNV}
\title{prepareCNV}
\usage{
prepareCNV(
  SeuratObject,
  gene.manifest,
  assay = "Spatial",
  ref.object = NULL,
  species = "human",
  genome = "hg19"
)
}
\arguments{
\item{SeuratObject}{A seurat object}

\item{gene.manifest}{A dataframe of gene information}

\item{assay}{Assay}

\item{species}{Species}

\item{genome}{Genome, hg38, hg19 or mm10}

\item{ref.data}{The reference seurat object data}
}
\value{
A list
}
\description{
Get common genes in 'expr.data' and 'ref.data'
}
