% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{regionprops}
\alias{regionprops}
\title{regionprops}
\usage{
regionprops(
  SeuratObject,
  cluster,
  region.select.method = "threshold",
  region.threshold = 20,
  cluster.info = "seurat_clusters",
  slice = "slice1"
)
}
\arguments{
\item{SeuratObject}{A Seurat object}

\item{cluster}{The cluster to analyze}

\item{region.threshold}{Only domains whose area are larger than `region.threshold` are remained.}

\item{cluster.info}{The label to store cluster information in `SeuratObject@meta.data`}

\item{slice}{}
}
\value{
A list of connected domains with information of barcodes and clusters
}
\description{
find all the connected domains of `cluster`
}
