% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{TLSAnalysis}
\alias{TLSAnalysis}
\title{TLSAnalysis}
\usage{
TLSAnalysis(SeuratObject, savePath = NULL, geneSets = NULL, ...)
}
\arguments{
\item{SeuratObject}{A Seurat object}

\item{savePath}{}
}
\value{
A Seurat object after NMF
}
\description{
calculate the score of TLS
}
